PATCH enable connection pool timeouts to be sub-second (#2)
this patch fixes a regression introduced in `v6.0.20` that made it not possible to setting connection pool timeouts below 1 second. 

the following patch:
- changes `ConnPoolConfig.TimeoutSeconds` to `float64`
- parses `-queryserver-config-query-pool-timeout` and `-queryserver-config-txpool-timeout` as `float64`
- add new flag `-queryserver-config-query-stream-pool-timeout` parsed as a `float64` for setting the the stream pool connection pool timeout

the patch is a subset of the following pull request https://github.com/vitessio/vitess/pull/6250/files which was opened after mstarr reported the behavior to sougou from PlanetScale. I've used the same flag name for `-queryserver-config-query-stream-pool-timeout` utilized in the PR which should make for an easy merge. 

vitess maintains three connection pools http://et.sy/9vlndg7
- `OltpReadPool` is used for `SELECT`s when `OLTP` mode is in use
- `OlapReadPool` is used for `SELECT`s when `OLAP` mode is in use
- `TxPool` is used for `INSERT`s/`UPDATE`s and transactions when `OLTP` mode is in use

the `-queryserver-config-query-pool-timeout` flag maps to `OltpReadPool.TimeoutSeconds` and controls how long vttablet waits for a connection from the `OltpReadPool`

the `-queryserver-config-txpool-timeout` flag maps to `TxPool.TimeoutSeconds` and controls how long vttablet waits for a connection from the `TxPool`

at the time of writing this in `v6.0.20` there is no equivalent flag for for the `OlapReadPool.TimeoutSeconds`, I also tested in `4.0.1` and the `-queryserver-config-query-pool-timeout` flag has no effect for OLAP mode. A look at the source code also confirms it never applied http://et.sy/1soa6nq

Signed-off-by: Mackenzie Starr <mstarr@etsy.com>