[PATCH] set_workload.patch
Adds the ability to toggle between OLAP (streaming) and OLTP modes on
the same connection.

This avoids opening an extra db connection and several round trips. We
did not attempt to upstream this as writes in OLAP mode are on
PlanetScale’s roadmap.

Note: the v8.0.0-rc1 release notes reference a fix to allow toggling
between OLAP and OLTP modes:
https://github.com/vitessio/vitess/pull/6691 but I confirmed with
Harshit Gangal from PlanetScale that this only applies when shard
targeting is not being used. As such, I've only ported our patch to
apply to the vtgate resolver which is used for executing stream
queries when shard targeting is in use

Signed-off-by: mstarr <mstarr@etsy.com>