[PATCH] mysql_hostname.patch
Overrides the advertised hostname of vttablet to the hostname of the
node instead of the `-db_host` flag.

We configure our tablets with `-db_host 127.0.0.1`. After upgrading to
4.x, we noticed the behavior of `TabletExternallyReparented X` changed
slightly in that it seems to issue a `CHANGE MASTER TO
MASTER_HOST=’127.0.0.1’, ...` on X's replica which fails. To avoid
this, our patch sets the advertised hostname explicitly to the
hostname of the node.

Thanks to @enisoc for the suggestion https://etsy.slack.com/archives/CPB2DN7UK/p1579632007000200

A proper patch might make separate config values for advertised and internal mysqld hostnames.

Signed-off-by: mstarr <mstarr@etsy.com>