[DAP-5266] multisharded vindex
Creates new multicolumn unique vindex requried for Etsy's Vitess-managed sharding strategy. The MultiSharded vindex accepts two columns representing a type_id and an id, and provides a way to specify a mapping between a type id and a vindex type. It then chooses the vindex that is used based on the type_id column.

This patch also adds logic that saves a mapping of Hybrid vindex instance names to Hybrid vindex instances, which are used by the MultiSharded vindex.