[DAP-5266] Add multisharded multicolumn vindex (#19)
* [DAP-5266] multisharded vindex

Creates new multicolumn unique vindex requried for Etsy's Vitess-managed sharding strategy. The MultiSharded vindex accepts two columns representing a type_id and an id, and provides a way to specify a mapping between a type id and a vindex type. It then chooses the vindex that is used based on the type_id column.

This patch also adds logic that saves a mapping of Hybrid vindex instance names to Hybrid vindex instances, which are used by the MultiSharded vindex.

* Clean up unit test; rename variables for consistency

* Add test for populating hybridVindexes

* Add test cases to check that vindex errors when either column is negative and that vindex supports string column values. Update vindex type name to 'etsy_multisharded_hybrid'. Error when vindex name can't be found in hybridVindexes map. Additional comments for clarity.

* Update input name from 'owner_type_to_vindex' to 'type_id_to_vindex'